+++ docs/changelog.md
@@ -40,6 +40,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: treat all votes that are not a special card as assignable (#102)
 * feat: evaluate need to talk flag for all kind of cards (#102)
 * feat: persist estimate based on custom field type (#102)
+* feat: only show supported custom fields in configuration (#102)
 
 ### [4.10.0] - 2020-03-28